:truck: api: move jury certification repository to src
