+++ api/src/certification/session-management/infrastructure/repositories/jury-certification-repository.js
@@ -1,11 +1,11 @@
 import _ from 'lodash';
 
-import { knex } from '../../../db/knex-database-connection.js';
-import { JuryCertification } from '../../../src/certification/session-management/domain/models/JuryCertification.js';
-import { ComplementaryCertificationCourseResultForJuryCertification } from '../../../src/certification/session-management/domain/read-models/ComplementaryCertificationCourseResultForJuryCertification.js';
-import { ComplementaryCertificationCourseResultForJuryCertificationWithExternal } from '../../../src/certification/session-management/domain/read-models/ComplementaryCertificationCourseResultForJuryCertificationWithExternal.js';
-import { CertificationIssueReport } from '../../../src/certification/shared/domain/models/CertificationIssueReport.js';
-import { NotFoundError } from '../../../src/shared/domain/errors.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
+import { NotFoundError } from '../../../../shared/domain/errors.js';
+import { CertificationIssueReport } from '../../../shared/domain/models/CertificationIssueReport.js';
+import { JuryCertification } from '../../domain/models/JuryCertification.js';
+import { ComplementaryCertificationCourseResultForJuryCertification } from '../../domain/read-models/ComplementaryCertificationCourseResultForJuryCertification.js';
+import { ComplementaryCertificationCourseResultForJuryCertificationWithExternal } from '../../domain/read-models/ComplementaryCertificationCourseResultForJuryCertificationWithExternal.js';
 
 const get = async function (certificationCourseId) {
   const juryCertificationDTO = await _selectJuryCertifications()
